requirements.txt
streamlit==1.32.0
pandas==2.2.0
psycopg2-binary==2.9.9
geopandas==0.14.3
folium==0.15.1
streamlit-folium==0.18.0
python-dotenv==1.0.1
streamlit-authenticator==0.3.2 